requirements.txt
flask
gunicorn
psycopg2-binary
python-dotenv
requests
flask-login
beautifulsoup4
PyPDF2
